refactor: Update weather folder name to discovery
