+++ src/entities/discovery/model/constants.ts
