+++ src/entities/discovery/model/index.ts
