+++ src/entities/discovery/model/model.d.ts
