+++ src/widgets/discovery/model/DUMMY.tsx
