+++ src/widgets/discovery/ui/discovery-card.tsx
@@ -1,4 +1,4 @@
-import type { Discovery } from '@/entities/weather';
+import type { Discovery } from '@/entities/discovery';
 import { cn } from '@/shared/lib';
 import WeatherIcon from './weather-icon';
 import WeeklyWeather from './weekly-weather';
