+++ src/widgets/discovery/ui/summary.tsx
